düzeltmeler

+++ backend/src/main/java/com/example/backend/services/FilesStorageServiceImpl.java
@@ -252,40 +252,44 @@ public class FilesStorageServiceImpl implements FileStorageService {
         Path directory = admin.resolve(title);
         Path userDirectory= user.resolve(title);
 
-        try {
-            Files.walk(directory)
-                    .sorted(Comparator.reverseOrder()) // Dosyaları ve dizinleri önce içten dışa doğru sıralar
-                    .forEach(path -> {
-                        try {
-                            Files.delete(path);
-                        } catch (IOException e) {
-                            throw new RuntimeException("Failed to delete file: " + path, e);
-                        }
-                    });
-        } catch (NoSuchFileException e) {
-            throw new RuntimeException("No such directory: " + directory, e);
-        } catch (DirectoryNotEmptyException e) {
-            throw new RuntimeException("Directory is not empty: " + directory, e);
-        } catch (IOException e) {
-            throw new RuntimeException("Failed to delete directory: " + directory, e);
+        if(Files.exists(directory) && Files.exists(userDirectory)){
+            try {
+                Files.walk(directory)
+                        .sorted(Comparator.reverseOrder()) // Dosyaları ve dizinleri önce içten dışa doğru sıralar
+                        .forEach(path -> {
+                            try {
+                                Files.delete(path);
+                            } catch (IOException e) {
+                                throw new RuntimeException("Failed to delete file: " + path, e);
+                            }
+                        });
+            } catch (NoSuchFileException e) {
+                throw new RuntimeException("No such directory: " + directory, e);
+            } catch (DirectoryNotEmptyException e) {
+                throw new RuntimeException("Directory is not empty: " + directory, e);
+            } catch (IOException e) {
+                throw new RuntimeException("Failed to delete directory: " + directory, e);
+            }
+            try {
+                Files.walk(userDirectory)
+                        .sorted(Comparator.reverseOrder()) // Dosyaları ve dizinleri önce içten dışa doğru sıralar
+                        .forEach(path -> {
+                            try {
+                                Files.delete(path);
+                            } catch (IOException e) {
+                                throw new RuntimeException("Failed to delete file: " + path, e);
+                            }
+                        });
+            } catch (NoSuchFileException e) {
+                throw new RuntimeException("No such directory: " + directory, e);
+            } catch (DirectoryNotEmptyException e) {
+                throw new RuntimeException("Directory is not empty: " + directory, e);
+            } catch (IOException e) {
+                throw new RuntimeException("Failed to delete directory: " + directory, e);
+            }
         }
-        try {
-            Files.walk(userDirectory)
-                    .sorted(Comparator.reverseOrder()) // Dosyaları ve dizinleri önce içten dışa doğru sıralar
-                    .forEach(path -> {
-                        try {
-                            Files.delete(path);
-                        } catch (IOException e) {
-                            throw new RuntimeException("Failed to delete file: " + path, e);
-                        }
-                    });
-        } catch (NoSuchFileException e) {
-            throw new RuntimeException("No such directory: " + directory, e);
-        } catch (DirectoryNotEmptyException e) {
-            throw new RuntimeException("Directory is not empty: " + directory, e);
-        } catch (IOException e) {
-            throw new RuntimeException("Failed to delete directory: " + directory, e);
-        }
+
     }
+
 
 }
